+++ ElvUI_MerathilisUI/Mainline/Modules/Bags/Core.lua
@@ -9,7 +9,6 @@ local B = E:GetModule('Bags')
 local _G = _G
 local strmatch, unpack, ceil = string.match, unpack, math.ceil
 
-local LE_ITEM_CLASS_CONTAINER = LE_ITEM_CLASS_CONTAINER
 local C_NewItems_IsNewItem, C_NewItems_RemoveNewItem, C_Timer_After = C_NewItems.IsNewItem, C_NewItems.RemoveNewItem, C_Timer.After
 local C_AzeriteEmpoweredItem_IsAzeriteEmpoweredItemByID = C_AzeriteEmpoweredItem.IsAzeriteEmpoweredItemByID
 local C_Soulbinds_IsItemConduitByItemInfo = C_Soulbinds.IsItemConduitByItemInfo
@@ -962,7 +961,7 @@ function module:Initialize()
 
 	function Backpack:OnInit()
 		AddNewContainer("Bag", 6, "BagReagent", filters.onlyBagReagent)
-		AddNewContainer("Bag", 16, "Junk", filters.bagsJunk)
+		AddNewContainer("Bag", 17, "Junk", filters.bagsJunk)
 		for i = 1, 5 do
 			AddNewContainer("Bag", i, "BagCustom" .. i, filters["bagCustom" .. i])
 		end
@@ -970,11 +969,12 @@ function module:Initialize()
 		AddNewContainer("Bag", 7, "AzeriteItem", filters.bagAzeriteItem)
 		AddNewContainer("Bag", 8, "Equipment", filters.bagEquipment)
 		AddNewContainer("Bag", 10, "BagCollection", filters.bagCollection)
-		AddNewContainer("Bag", 14, "Consumable", filters.bagConsumable)
+		AddNewContainer("Bag", 15, "Consumable", filters.bagConsumable)
 		AddNewContainer("Bag", 11, "BagGoods", filters.bagGoods)
-		AddNewContainer("Bag", 15, "BagQuest", filters.bagQuest)
+		AddNewContainer("Bag", 16, "BagQuest", filters.bagQuest)
 		AddNewContainer("Bag", 12, "BagAnima", filters.bagAnima)
 		AddNewContainer("Bag", 13, "BagRelic", filters.bagRelic)
+		AddNewContainer("Bag", 14, "BagStone", filters.bagStone)
 
 		f.main = MyContainer:New("Bag", { Bags = "bags", BagType = "Bag" })
 		f.main.__anchor = { "BOTTOMRIGHT", -4, 50 }
@@ -1120,7 +1120,7 @@ function module:Initialize()
 	end
 
 	local function isItemNeedsLevel(item)
-		return item.link and item.quality > 1 and module:IsItemHasLevel(item)
+		return item.link and item.quality > 1 and (module:IsItemHasLevel(item) or item.classID == Enum.ItemClass.Gem)
 	end
 
 	local function isItemExist(item)
@@ -1408,6 +1408,8 @@ function module:Initialize()
 			label = GetCustomGroupTitle(settings.Index)
 		elseif name == "BagReagent" then
 			label = L["Reagent Bag"]
+		elseif name == "BagStone" then
+			label = GetSpellInfo(404861)
 		end
 		if label then
 			self.label = self:CreateFontString(nil, "ARTWORK")