+++ ElvUI_MerathilisUI/Mainline/Modules/Skins/AddOns/Blizzard_ItemSocketingUI.lua
@@ -14,30 +14,9 @@ local function LoadSkin()
 	ItemSocketingFrame:Styling()
 	module:CreateBackdropShadow(ItemSocketingFrame)
 
-	ItemSocketingFrame:DisableDrawLayer("ARTWORK")
-
 	local title = select(18, ItemSocketingFrame:GetRegions())
 	title:ClearAllPoints()
 	title:SetPoint("TOP", 0, -5)
-
-	for i = 1, _G.MAX_NUM_SOCKETS do
-		local bu = _G["ItemSocketingSocket"..i]
-		local shine = _G["ItemSocketingSocket"..i.."Shine"]
-
-		_G["ItemSocketingSocket"..i.."BracketFrame"]:Hide()
-		_G["ItemSocketingSocket"..i.."Background"]:SetAlpha(0)
-		select(2, bu:GetRegions()):Hide()
-
-		bu:SetPushedTexture("")
-		bu:GetHighlightTexture():SetColorTexture(1, 1, 1, .25)
-		bu.icon:SetTexCoord(unpack(E.TexCoords))
-
-		shine:ClearAllPoints()
-		shine:SetPoint("TOPLEFT", bu)
-		shine:SetPoint("BOTTOMRIGHT", bu, 1, 0)
-
-		bu.bg = module:CreateBDFrame(bu, .25)
-	end
 end
 
 S:AddCallbackForAddon("Blizzard_ItemSocketingUI", LoadSkin)