+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ### Changes ###
 
-  * Chat: Fix a non Retail Chat error.. narf
+  * Chat: Added an updated version for the Online/Offline. Credits fang2hou
+  * Chat: Rearrange Chat settings
+  * Tooltip: Added icons for Achievements
+  * FriendList: Fixed an error with name color
+  * Skins: Updated Auctionator skin
+  * Skins: Updated ItemSocket/ProfessionCustomOrder skins
+  * Bagss: Added an own category for Primodial Stones